added ignoredOptions to IndicesSelectComboBox
Updated the IndicesSelectComboBox to take an optional ignoredOptions
prop. This prop will take a list of indice names to ignore from the
search results and not display as options. This is specifically to not
allow selecting indices in the Add indices flyout that are already a
part of the search application.