Enterprise Search: added ignoredOptions to IndicesSelectComboBox (#157065)
## Summary

Updated the IndicesSelectComboBox to take an optional ignoredOptions
prop. This prop will take a list of indice names to ignore from the
search results and not display as options. This is specifically to not
allow selecting indices in the Add indices flyout that are already a
part of the search application.

### Screenshots


https://user-images.githubusercontent.com/1972968/236926701-fcae8aa9-a396-487c-b345-1a50a66d2dce.mov

Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>