Fix prop drilling issue and expose onRefreshChange (#157468)
## Summary

The PR fixes the below
1. `props.onRefresh` prop was missed during prop drilling. Due to this
the Auto Refresh feature is broken.
2. `props.onRefreshChange` event is currently not exposing a callback
function directly on the SearchBar. This is required on APM side as when
a user changes the Refresh Interval, we need to know the set duration
and paused toggle to update the application URL.

(cherry picked from commit 13bfd933d3bc1f889dc910c58ad6af0d5a0c586d)