[ci/jest_parallel] Reduce node process memory to 4096
Jest tests on CI are currently allocated 14336mb of memory.  This was
introduced when we were running suites in parallel, and left in place
after we reverted to running in band.

This lowers the memory limit to 4096.  We want to shift some of the
memory allocation to elasticsearch to allow larger integration tests.