[Profiling] Replace mget calls in the Stacktraces view (#157808)
## Summary

This PR removes all references to `mget` calls as started in
https://github.com/elastic/prodfiler/issues/2775.

I replaced the metadata lookup in the stacktraces view with the
Elasticsearch plugin endpoint (i.e. `_profiling/stacktraces).

The main consequence of this change is that we no longer need to cache
the results when decoding the key-value pairs. However, the endpoints
that clear the internal caches will need to be remove separately so that
we can remove the calls to these endpoints from other clients.

### For maintainers

- [x] This was checked for breaking API changes and was [labeled
appropriately](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)