[Infrastructure UI] Correct time generation Default views (#157739)
## 📓 Summary

This PR fixes generating a static `time` property for the default saved
view for inventory and metrics-explorer, switching to a new date
timestamp on each request as the previous behaviour.

## 🧪 Testing

Navigate to `Infrastructure -> Inventory`, remove all the URL query
params and check that the time selected is the current time for the
Default view.
The same behaviour is expected for the metrics explorer Default view.

---------

Co-authored-by: Marco Antonio Ghiani <marcoantonio.ghiani@elastic.co>
(cherry picked from commit 26a559d3d14b1c863cb18bbb8413ae5c67ff4f36)