Allow pagination of the risk score endpoint
We now accept an `after_keys` parameter which corresponds to the `after`
keys of the respective identifier aggregations involved. It is currently
up to the client to take the key from the response and pass it to the
next request.

Since the bucket size is not exposed as an http parameter for now,
whether or not to "paginate" depends on the presence/absence of
after_keys in the response.