serverless_search: remove enterprise_search plugin dep
The current design of the serverless search solution does not require
any enterprise_search pages. Given this we are removing it from being
enabled for now in favor of just the serverless_search plugin. we may
enable it again later, or move the views needed to another plugin that
can be shared between both enterprise_search and serverless_search