[Serverless/sidenav] Hide collapsed sidebar on smaller screen, move nav button to the header (#157999)
## Summary

Partially addresses https://github.com/elastic/kibana/issues/157819,
https://github.com/elastic/kibana/issues/157032

To address https://github.com/elastic/kibana/issues/157032 I've hidden
the collapsed side nav on a smaller screen. Since there is no more side
nav on smaller screen, I had to move the expand/collapse button to the
header. But it was also proposed to have that button in the header
constantly
[here](https://github.com/elastic/kibana/issues/157819#issuecomment-1551023516),
so I decided to just leave it there both for smaller and larger screens.
This makes us closer to the recent version (17/05) that
@MichaelMarcialis proposed _(Not sure if we it was already decide to go
that route?)_




https://github.com/elastic/kibana/assets/7784120/7a163572-ee0a-41d1-89df-2882b786be94