[Security Solution] Failed getFieldsForIndexPattern calls can result in Exception Flyout getting stuck in loading state (#158110)
Original ticket #158110

These changes fixes the issue with the exception flyout which can be stuck in loading state in case getFieldsForIndexPattern throws an exception.

Fixed by putting the getFieldsForIndexPattern call in try/catch. We use this call to fetch extended information about the fields to show warning to the user in case there are some index issues. If getFieldsForIndexPattern fails and throws an exception we will continue using fields without conflicts/unmapped information.

I also, noticed that we do not fetch extended information for the case where user uses index patterns instead of data views. Fixed this issue in useFetchIndex.