[Cases] Decode all responses from the cases client (#158084)
## Summary

The `encode` function of `io-ts` does not validate the input. For this
reason, this PR replaces all `encode` instances in the cases client with
`decode` which throws an error if the input is not as expected. In
combination with https://github.com/elastic/kibana/pull/156813 all
responses from the cases client will a) strip out excess fields and b)
throw an error if the response does not match the expected schema.

I did a change in the `data.json.gz` file. One of the create case user
actions had `null` for `new_value`. It is not possible to have this. I
put the minimum attributes (title, description, and tags) in the
`new_value` and zip the file again.

### For maintainers

- [x] This was checked for breaking API changes and was [labeled
appropriately](https://www.elastic.co/guide/en/kibana/master/contributing.html#kibana-release-notes-process)

---------

Co-authored-by: Jonathan Buttner <jonathan.buttner@elastic.co>
Co-authored-by: Jonathan Buttner <56361221+jonathan-buttner@users.noreply.github.com>