[ML] Stabilize and reenable notifications tests (#158546)
## Summary

This PR stabilizes ML notifications tests by introducing leniency when
checking for the notifications table in loading state. It also reenables
the test suite.

### Other changes

After the test cleanup was done, we still had a left-over saved object.
This was caused by the test creating a job for `space1` without actually
creating that space. In order to fix this, we now let the test properly
create and cleanup that space.