[data view field editor] Refactor business logic from hooks to controller (#156203)
## Summary

This PR moves a lot of business logic from `field_preview_context.tsx`
to the controller. No functionality is changed. This is a step to simply
the code and eventually reduce bugs.

My approach to this change was -
1. Move `useState` calls to controller state, adding necessary setters
2. Move `useCallback` calls to controller methods.
3. Move `useEffect` calls - this one is the most complex - they get
called when state values are set. Sometimes the code is inlined or a new
private method might be created.

The bits that remain are mostly related to form values referred to as
`params`. I'm saving this for a subsequent PR as this PR is large enough
as it is.

Closes https://github.com/elastic/kibana/issues/151243

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>