[Maps] Update asset tracking tutorial with production consideration for rule check interval (#158664)
Fixes #127289

## Summary

Updates the asset tracking tutorial docs with considerations for running
in production.

For a better demo experience the tutorial uses a 5 second interval for
rule checking, but a very low interval can be problematic for running in
production.

I'm suggesting backporting these doc changes to 8.3 which the first
release to have a **Circuit breakers** section in
https://www.elastic.co/guide/en/kibana/8.3/alerting-production-considerations.html
that is most relevant to the added note in this PR.

---------

Co-authored-by: gchaps <33642766+gchaps@users.noreply.github.com>
(cherry picked from commit 9d7d6f92133d100efee9eefd80fed9e097427f8d)