[Security Solution] Remove legacy navigation and related logic (#158094)
part of: https://github.com/elastic/kibana/issues/157847
closes: https://github.com/elastic/kibana/issues/145718

### Background

The new navigation became the default navigation for Security on 8.4. To
have a smooth transition we added an advanced setting to use the old
navigation.

Since then the legacy navigation has become outdated, there are links in
Security that are not accessible through it, such as the landing pages
for the sub-sections:
- `/security/dashboards`
- `/security/explore`
- `/security/manage`

With the introduction of the new Security AI design, more of those
landing pages that are not compatible with the legacy navigation design
will be added (e.g. Rules). And it was starting to become overcrowded
since there was no possibility to collapse groups.

Also, over time it has become harder and harder to maintain both
versions at the same time, all the new pages added to security were
having to duplicate the navigation configurations for both versions and
also test everything twice.

On top of that, the legacy navigation won't be supported on the Security
Serverless projects, everything will work with the new one exclusively.

## Docs

The Security documentation assumes the new navigation is used
everywhere, there's no mention of the old navigation, only one small
section about the advanced setting (which is actually outdated):

https://www.elastic.co/guide/en/security/8.7/advanced-settings.html#_enable_grouped_navigation
Which will need to be removed for 8.9.

## Summary

Cleans the old navigation and the advanced setting
(`securitySolution:enableGroupedNav`) to turn it on. Removes the
telemetry (which shows almost no usage of it) and external dependencies
as well.

Only the new navigation will be available. 
All links should work only with the `app_links` architecture from now
on.

Old Nav ❌ 


![old_nav](https://github.com/elastic/kibana/assets/17747913/b95ff48c-73d4-45f6-93e5-8bec3fb24052)


New nav ✅ 


![new_nav](https://github.com/elastic/kibana/assets/17747913/8695c63e-df97-4ffc-b504-b4f48a366b3b)

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>