[Enterprise Search] Optimize refresh usage (#158834)
## Summary

This optimizes the usage of `refresh` across the Enterprise Search
plugin.
1) Removed most usages of refresh, where the document is not expected to
be accessed via search requests (or up-to-date-ness in the next search
request is not essential).
2) Changed `refresh: true` to `refresh: 'wait_for'` for adding a
connector document. We need to wait for this refresh to complete, or the
redirect after adding a new connector will show a broken page.