serverless_search: remove fallback cloud id
Updated the API key panel for the getting started page to not render a
fallback cloud id when its unavailable. In the future this cloud id
should always be set for a serverless project, but defaulting to a cloud
id was confusing in the QA environment since it was incorrect.