[Security Solution] Remove extra data structures in create/edit rule forms (#157749)
## Summary

This PR refactors the forms used by the Detection Engine create and edit
rule pages. The existing forms have a lot of complexity in how data is
managed: each of the 4 steps (Define, About, Schedule, and Actions) map
to a separate component, and these individual components own the form
for that step. However, some steps have dependencies on data in other
steps. In addition, Rule Preview and submitting the rule to the API
require the data from all steps. To resolve this, I moved the form
ownership from the individual step components to the Create/Edit Rule
page components and passed the appropriate form into each individual
step component.

With the forms owned at the Page level, many of the additional data
structures we previously needed to pass form data around are no longer
necessary.
#### Removed
- `formHooks`, `setFormHooks` - These hooks were used to let step
components pass the form's `validate` function back up to the Page
component. With the forms created in the Page component, the `validate`
functions are available without additional structures.
- `stepsData`, `setStepsData` - This data structure held the form data
from each step and the validation status. We now use the live form data
through `useFormData` without additional sidecar data structures.
- `defineRuleData`/`aboutRuleData`/`scheduleRuleData` - Sidecar data
structures for rule preview, also held the form data from each step. We
now use the live form data through `useFormData` without additional
sidecar data structures.

#### Simplified
- `editStep()` no longer relies on `formHooks`
- `submitStep()` no longer relies on `formHooks`
- Step forms no longer need `watch` and `onChange` configurations

### Key Implementation Notes
In order for `useFormData` to be reliable, the forms for all steps must
stay in the DOM at all times while on the Create/Edit pages. Thus the
"read only" view has been separated from the editable view, and the
visibility of each view is toggled by hiding elements with CSS rather
than removing them from the DOM.

If the form were to be removed from the DOM when switching to a read
only view, the values from `useFormData` would reset to the default
values so we would need a sidecar object to hold the form values when
the form is not rendered. It would also require re-loading the values
from the sidecar into the form when the form gets re-added to the DOM.
Keeping the form in the DOM at all times allows us to rely on
`useFormData` consistently, which simplifies the data model.

### Bug Fixes
- When moving back to step 1 from step 2 by clicking "Edit", data is no
longer lost in step 2

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>