Debounce changes to options list
Moves `selectedOptions` and `existsSelected` properties to component
state. The changes are debounced for a split second prior to updating
embeddable state to provide a smoother experience when selecting or
deselecting options.

Separating these properties to componentState should also make it easier
to have an ["Apply" button](https://github.com/elastic/kibana/issues/135459) in a future PR.