[Logs UI] Fix Log Categorization UI failure due to infinite loop (#159090)
## 📓 Summary

With the usage of the `logViewReference` object (instead of the
primitive `sourceId`) as a dependency for the log entries categories
data fetching, the `useTrackedPromise` was continuously recreating the
`getLogEntryCategoryDatasets` and `getTopLogEntryCategories` fetchers,
which were consequently causing an infinite loop when called inside the
related side effect and used as dependency.

Using the `logViewReference.logViewId` property (equivalent to the
legacy `sourceId` string) restores the expected behaviour.

Co-authored-by: Marco Antonio Ghiani <marcoantonio.ghiani@elastic.co>
(cherry picked from commit d6105d0bfe0634a0543e52ce3c1275ad0bd64250)